pyproject.toml
[tool.poetry]
name = "django-dynamic-storage"
version = "0.1.1"
description = "Use and choose storages at runtime based on your logic"
authors = ["Amir Khalife <eng.amir.bu@gmail.com>"]
license = "Apache-2.0"
packages = [{include = "dynamic_storage"}]

[tool.poetry.dependencies]
python = ">=3.8"

[tool.poetry.group.test.dependencies]
Django = "^4.2.1"
pillow = "^9.5.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
django_find_project = false
pythonpath = "tests"
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.isort]
known_first_party="django,pytest"
known_apps="dynamic_storage"
lines_after_imports=2
profile = "black"
py_version=39
sections="FUTURE,STDLIB,FIRSTPARTY,THIRDPARTY,APPS,LOCALFOLDER"

